atualizacao-atividade-10

+++ bike-rent-poo-master/src/app.ts
@@ -7,9 +7,12 @@ import { BikeNotFoundError } from "./errors/bike-not-found-error";
 import { UnavailableBikeError } from "./errors/unavailable-bike-error";
 import { UserNotFoundError } from "./errors/user-not-found-error";
 import { DuplicateUserError } from "./errors/duplicate-user-error";
+import { OpenRentsError } from "./errors/open-rents-error";
 import { RentRepo } from "./ports/rent-repo";
 import { UserRepo } from "./ports/user-repo";
 import { BikeRepo } from "./ports/bike-repo";
+import { RentError } from "./errors/rent-error";
+
 
 export class App {
     crypt: Crypt = new Crypt()
@@ -45,9 +48,15 @@ export class App {
     }
 
     async removeUser(email: string): Promise<void> {
-        await this.findUser(email)
-        await this.userRepo.remove(email)
-    }
+        const user = await this.findUser(email);
+        const openRents = await this.rentRepo.findOpenRentsFor(email);
+      
+        if (openRents.length > 0) {
+          throw new OpenRentsError(); // Lançar uma exceção se houver aluguéis em aberto.
+        }
+      
+        await this.userRepo.remove(email);
+      }
     
     async rentBike(bikeId: string, userEmail: string): Promise<string> {
         const bike = await this.findBike(bikeId)
@@ -64,7 +73,7 @@ export class App {
     async returnBike(bikeId: string, userEmail: string): Promise<number> {
         const now = new Date()
         const rent = await this.rentRepo.findOpen(bikeId, userEmail)
-        if (!rent) throw new Error('Rent not found.')
+        if (!rent) throw new RentError()
         rent.end = now
         await this.rentRepo.update(rent.id, rent)
         rent.bike.available = true
